% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_clustering.R
\name{Clustering}
\alias{Clustering}
\title{Clustering function}
\usage{
Clustering(data_clustering, nb_clust_possible, n_iter = 5, Dmin = 0.28)
}
\arguments{
\item{data_clustering}{dataset with Contrast and SigStren for each individuals (as SampleName) and each markers (as MarkerName)}

\item{nb_clust_possible}{number of cluster possible (ploidy+1)}

\item{n_iter}{number of iterations to perform for clustering}

\item{Dmin}{minimal distance between two clusters}
}
\value{
list of results of clustering
}
\description{
Clustering function to run clustering with no parallelization process nor auto save
}
\examples{
\dontrun{
file.create = system.file("extdata/output_create/1","1_1_to_clust.Rdata", package = "GenoTriplo")
load(file.create)
ploidy=3
Clustering(data_clustering=data_clustering,nb_clust_possible=ploidy+1)
}


}
