% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_run_clustering.R
\name{Clustering_from_dir}
\alias{Clustering_from_dir}
\title{Clustering function from directory}
\usage{
Clustering_from_dir(
  dir_from_create,
  ploidy,
  n_iter = 5,
  Dmin = 0.28,
  new.dir = NULL
)
}
\arguments{
\item{dir_from_create}{directory of the file to clusterize.}

\item{ploidy}{ploidy of individuals}

\item{n_iter}{number of iterations to perform for clustering}

\item{Dmin}{minimal distance between two clusters}

\item{new.dir}{new directory name for saving in './output_clustering'}
}
\value{
Autosave of the results in new.dir
}
\description{
Clustering function to run clustering with no parallelization process and autosave
}
\examples{
\dontrun{
dir.create = system.file("extdata/output_create/1", package = "GenoTriplo")
ploidy = 3
Clustering_from_dir(dir_from_create=dir.create,ploidy=ploidy)
}
}
