% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_clustering.R
\name{Clustering_from_file}
\alias{Clustering_from_file}
\title{Clustering function from file}
\usage{
Clustering_from_file(
  dirname,
  filename,
  ploidy,
  n_iter = 5,
  Dmin = 0.28,
  new.dir = NULL
)
}
\arguments{
\item{dirname}{directory of the file to clusterize.}

\item{filename}{filename of the data to clusterize}

\item{ploidy}{ploidy of individuals}

\item{n_iter}{number of iterations to perform for clustering}

\item{Dmin}{minimal distance between two clusters}

\item{new.dir}{new directory name for saving}
}
\value{
Autosave of the results
}
\description{
Clustering function to run clustering with no parallelization process and autosave
Used by Clustering_from_dir. Shouldn't be used else.
}
\examples{
\dontrun{
dir.create = system.file("extdata/output_create/1", package = "GenoTriplo")
file.create = system.file("extdata/output_create/1","1_1_to_clust.Rdata", package = "GenoTriplo")
ploidy=3
Clustering_from_file(dirname=dir.create,filename=file.create,ploidy=ploidy)
}

}
