% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_run_clustering.R
\name{Clustering_parallele_from_dir}
\alias{Clustering_parallele_from_dir}
\title{Launch parallel clustering from directory}
\usage{
Clustering_parallele_from_dir(
  dir_from_create,
  new.dir = NULL,
  ploidy,
  n_iter = 5,
  Dmin = 0.28,
  n_core = 1
)
}
\arguments{
\item{dir_from_create}{directory name with dataset files saved in './output_create'}

\item{new.dir}{directory name where output is saved as in './output_clustering'}

\item{ploidy}{ploidy of individuals}

\item{n_iter}{number of iterations of clustering}

\item{Dmin}{threshold distance between two clusters}

\item{n_core}{number of cores used for parallelization}
}
\value{
the result of clustering or automatically save a list of objects if a saving name has been provided
}
\description{
Launch the clustering phase in parallel from the dataset with SampleName, Contrast and SigStren for each markers (MarkerName).
}
\examples{
\dontrun{
dir.create = system.file("extdata/output_create/1", package = "GenoTriplo")
Clustering_parallele_from_dir(dir_from_create=dir.create,ploidy=3)
}

}
