% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_create_dataset.R
\name{Create_Dataset}
\alias{Create_Dataset}
\title{Create dataset in appropriate format}
\usage{
Create_Dataset(data, save_name = NULL)
}
\arguments{
\item{data}{dataframe with probeset_id as first variable (markername finishing by -A or -B depending on the probeset) and individuals as variable with luminescence values for each probeset (dataset created by bash code by shiny app)}

\item{save_name}{saving name}
}
\value{
number of individuals and markers (automatically save the dataset)
}
\description{
Create SigStren and Contrast variables from luminescence values of probeset A and B of each markers and return a dataframe to be used for clustering or save the result if a saving name is given
}
