% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_create_dataset.R
\name{Create_Dataset_from_file}
\alias{Create_Dataset_from_file}
\title{Create dataset from filename in appropriate format}
\usage{
Create_Dataset_from_file(
  filename = "./AxiomGT1.summary.txt",
  dir.name = NULL,
  indiv_name = NULL,
  marker_name = NULL,
  n_marker_chunk = 4000,
  simplify.names = TRUE
)
}
\arguments{
\item{filename}{filename from AXAS : \code{AxiomGT1.summary.txt}}

\item{dir.name}{saving name (the number of the chunk and _to_clust is automatically added)}

\item{indiv_name}{vector with name of individuals to keep}

\item{marker_name}{vector with name of marker to keep}

\item{n_marker_chunk}{nombre de lignes par chunk (nombre pair car chaque marker est par pair de ligne)}

\item{simplify.names}{Samplenames might be like so : IndivXX_a1_a1.CEL (if from Axiom) -> allows to select only IndivXX as SampleName}
}
\value{
save file in \code{./output_create/dir.name} as \code{k_to_clust.Rdata} with k the chunk number.
}
\description{
Create SigStren and Contrast variables from luminescence values of probeset A and B of each markers and return a dataframe to be used for clustering or save the result if a saving name is given
}
\examples{
\dontrun{
filepath = system.file("extdata", "AxiomGT1.summary_examples.txt", package = "GenoTriplo")
Create_Dataset_from_file(filename=filepath)
}
}
