% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_create_genofiles.R
\name{create_genofiles}
\alias{create_genofiles}
\title{Create files of genotype and indiv/marker stats from genotype .Rdata files}
\usage{
create_genofiles(dir.name.save, new.dir, ploidy, corres_ATCG = NULL)
}
\arguments{
\item{dir.name.save}{directory with .Rdata files with genotype}

\item{new.dir}{new directory for saving files}

\item{ploidy}{ploidy of the population}

\item{corres_ATCG}{dataframe with correspondance between Allele_A/Allele_B to ATCG}
}
\value{
names of the indivCR and markerCR files
}
\description{
Create files of genotype and indiv/marker stats from genotype .Rdata files
}
\examples{
\dontrun{
ploidy = 3
dir.name.save = system.file("extdata/output_genotyping/1", package = "GenoTriplo")
new.dir = "1"
create_genofiles(dir.name.save=dir.name.save,new.dir=new.dir,ploidy=ploidy)
}

}
