% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoTriplo_to_clust.R
\docType{data}
\name{GenoTriplo_to_clust}
\alias{GenoTriplo_to_clust}
\title{Example of dataset for clustering}
\format{
A dataframe with 500 rows (corresponding to an individual for a given marker) and 4 columns (SigStren,Contrast,SampleName,MarkerName)
}
\usage{
GenoTriplo_to_clust
}
\description{
Example of dataset for clustering
}
\keyword{datasets}
