% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoTriplo_to_geno.R
\docType{data}
\name{GenoTriplo_to_geno}
\alias{GenoTriplo_to_geno}
\title{Example of dataset for genotyping}
\format{
A list of 10 each element being the result of clustering for a given marker
}
\usage{
GenoTriplo_to_geno
}
\description{
Example of dataset for genotyping
}
\keyword{datasets}
