% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_run_genotyping.R
\name{Genotyping_from_file}
\alias{Genotyping_from_file}
\title{Run genotyping from file}
\usage{
Genotyping_from_file(
  dirname,
  filename,
  new.dir = NULL,
  ploidy,
  SeuilNoCall = 0.85,
  SeuilNbSD = 2.8,
  SeuilSD = 0.28,
  cr_marker = 0.97,
  fld_marker = 3.4,
  hetso_marker = -0.3,
  same.pop = TRUE,
  tryEqHW = TRUE
)
}
\arguments{
\item{dirname}{name of the directory where the file is}

\item{filename}{name of the file with the data for genotyping}

\item{new.dir}{new directory name for saving}

\item{ploidy}{ploidy of individuals}

\item{SeuilNoCall}{threshold of the probability of belonging to a cluster}

\item{SeuilNbSD}{threshold for the distance between an individuals and his cluster (x=Contrast)}

\item{SeuilSD}{threshold for the standard deviation of a cluster (\code{SeuilSD*(1+0.5*abs(mean_contrast_cluster))})}

\item{cr_marker}{call rate threshold}

\item{fld_marker}{FLD threshold}

\item{hetso_marker}{HetSO threshold}

\item{same.pop}{Boolean : are individuals from a same population}

\item{tryEqHW}{Only if same.pop==TRUE else ignore. If TRUE, will delete cluster that shouldn't happen under H-W equilibrium (heterozygote with less individuals than both homozygote in triploid for example). Additionnaly, the cluster should contain less than 5\% of the individuals. Else it stays.}
}
\value{
Void. Save the results
}
\description{
Run genotyping from file
}
\examples{
\dontrun{
ploidy=3
dir_from_clustering=system.file("extdata/output_clustering/1", package = "GenoTriplo")
file_from_clustering=system.file(
                      "extdata/output_clustering/1",
                      "1_1_to_geno.Rdata", package = "GenoTriplo")
Genotyping_from_file(dirname=dir_from_clustering,ploidy=ploidy,filename=file_from_clustering)
}
}
