% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_run_genotyping.R
\name{Genotyping_parallele_from_dir}
\alias{Genotyping_parallele_from_dir}
\title{Launch genotyping phase in parallel from directory}
\usage{
Genotyping_parallele_from_dir(
  dir_from_clustering,
  new.dir = NULL,
  ploidy,
  SeuilNoCall = 0.85,
  SeuilNbSD = 2.8,
  SeuilSD = 0.28,
  n_core = 1,
  corres_ATCG = NULL,
  same.pop = TRUE,
  cr_marker = 0.97,
  fld_marker = 3.4,
  hetso_marker = -0.3,
  tryEqHW = TRUE
)
}
\arguments{
\item{dir_from_clustering}{directory with files from clustering phase}

\item{new.dir}{name of the new directory for saving results in './output_genotyping'}

\item{ploidy}{ploidy of individuals}

\item{SeuilNoCall}{threshold of the probability of belonging to a cluster}

\item{SeuilNbSD}{threshold for the distance between an individuals and his cluster (x=Contrast)}

\item{SeuilSD}{threshold for the standard deviation of a cluster (SeuilSD*(1+0.5*abs(mean_contrast_cluster)))}

\item{n_core}{number of cores used for parallelization}

\item{corres_ATCG}{dataframe with the correspondence between A/B of AXAS and A/T/C/G (three columns : probeset_id, Allele_A, Allele_B)}

\item{same.pop}{Boolean : are individuals from a same population}

\item{cr_marker}{call rate threshold}

\item{fld_marker}{FLD threshold}

\item{hetso_marker}{HetSO threshold}

\item{tryEqHW}{Only if same.pop==TRUE else ignore. If TRUE, will delete cluster that shouldn't happen under H-W equilibrium (heterozygote with less individuals than both homozygote in triploid for example). Additionnaly, the cluster should contain less than 5\% of the individuals. Else it stays.}
}
\value{
Void. Save the results.
}
\description{
Function that launch the genotyping phase from the dataset with SampleName, Contrast and SigStren for each markers and the result of the 'Run_clustering' function.
}
\examples{
\dontrun{
ploidy=3
dir_from_clustering=system.file("extdata/output_clustering/1", package = "GenoTriplo")
Genotyping_parallele_from_dir(dir_from_clustering=dir_from_clustering,ploidy=ploidy)
}

}
