% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App_GenoTriploShiny.R
\name{launch_GenoShiny}
\alias{launch_GenoShiny}
\title{Shiny App for genotyping}
\usage{
launch_GenoShiny()
}
\value{
void : most results are automatically saved
}
\description{
Launch a shiny interface to use GenoTriplo. Really easy to use and user friendly, this will help you gain time !
}
